Refactor move menu_card_view to view_component dir.
